+++ src/main/java/domain/board/BoardPosition.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.board;
 
 public record BoardPosition(
         int x,
