+++ src/test/java/domain/JanggiTest.java
@@ -2,6 +2,10 @@ package domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import domain.board.Board;
+import domain.board.BoardPosition;
+import domain.piece.Piece;
+import domain.piece.PieceType;
 import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
