+++ src/test/java/domain/OffsetTest.java
@@ -1,10 +1,11 @@
 package domain;
 
-import static domain.Offset.*;
+import static domain.board.Offset.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
+import domain.board.Offset;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
